Map_Merging/requirements.txt
colorama==0.4.3
Django==3.2.3
h5py==3.1.0
imutils==0.5.4
loguru==0.5.3
matplotlib==3.1.2
numba==0.54.1
numpy==1.17.4
open3d==0.15.2
pandas==1.3.5
Pillow==9.1.0
progressbar33==2.4
PyYAML==6.0
scipy==1.8.0
tensorflow==2.8.0
tensorflow_gpu==2.7.0
tf_slim==1.1.0
